MarkUsed use case
